Group inputs related to ingredients
